Add explicit exact lookup name
During testing of the in lookup name for the hashid lookup type, it was
determined the explicit behaviors around the exact lookup type were also
not defined. In something like django-filters, this could make for odd
behaviors depending on how the project defines the `DEFAULT_LOOKUP_EXPR`
setting. It defaults to exact, but if set to something else then the
exact lookup may not behave as expected.